Refactored folder architecture
