+++ CMakeLists.txt
@@ -1,36 +1,39 @@
 cmake_minimum_required (VERSION 3.0)
 project (hop)
 
-#####################
-### Platform specific
-#####################
-if(UNIX)
-   if(APPLE)
-        set( CMAKE_OSX_DEPLOYMENT_TARGET 10.11 )
-        set( NOC_IMPL_DEFINE "NOC_FILE_DIALOG_OSX" )
-   else()
-        set( HOP_LIBRT_FLAG "-lrt" )
-        set( NOC_IMPL_DEFINE "NOC_FILE_DIALOG_GTK" )
-   endif()
-   set( CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra -pedantic -Wstrict-aliasing=2 " )
-else()
-   set( CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS}" )
-   set( HOP_WRAP_SRC wrap_library/hop_wrap_windows.cpp )
-   set( NOC_IMPL_DEFINE "NOC_FILE_DIALOG_WIN32" )
-endif()
-
-set( default_build_type "Release" )
-set( CMAKE_CXX_STANDARD 14 )
-set(ROOT_DIR ${CMAKE_SOURCE_DIR})
+set( CMAKE_VERBOSE_MAKEFILE on )
 
 # includes cmake/FindSDL2.cmake
 set(CMAKE_MODULE_PATH ${PROJECT_SOURCE_DIR}/cmake)
 find_package(SDL2 REQUIRED)
 find_package(OpenGL REQUIRED)
 
-set(CMAKE_CXX_FLAGS "-ffast-math -fno-exceptions -fno-rtti ${CMAKE_CXX_FLAGS}")
+set( default_build_type "Release" )
+set( CMAKE_CXX_STANDARD 14 )
+set(ROOT_DIR ${CMAKE_SOURCE_DIR})
 
-set( CMAKE_VERBOSE_MAKEFILE on )
+#####################
+### Platform specific
+#####################
+if(UNIX)
+   set( PLATFORM_LINK_FLAGS "-lpthread -ldl" )
+   set( CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra -pedantic -ffast-math -fno-exceptions -fno-rtti -Wstrict-aliasing=2 " )
+
+   if(APPLE)
+        set( CMAKE_OSX_DEPLOYMENT_TARGET 10.11 )
+        set( NOC_IMPL_DEFINE "NOC_FILE_DIALOG_OSX" )
+   else()
+        find_package( PkgConfig REQUIRED )
+        pkg_check_modules( GTK3 REQUIRED gtk+-3.0 )
+        set( PLATFORM_LINK_FLAGS "${PLATFORM_LINK_FLAGS} -lrt -l${GTK3_LIBRARIES}" )
+        set( NOC_IMPL_DEFINE "NOC_FILE_DIALOG_GTK" )
+        set( PLATFORM_EXTRA_INCLUDES ${GTK3_INCLUDE_DIRS} )
+   endif()
+else() # Windows
+   set( CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /fp:fast" )
+   set( HOP_WRAP_SRC wrap_library/hop_wrap_windows.cpp )
+   set( NOC_IMPL_DEFINE "NOC_FILE_DIALOG_WIN32" )
+endif()
 
 #####################
 ### Asan
@@ -42,12 +45,12 @@ if(SAN)
 endif()
 
 #####################
-### HOP common src
+### HOP common
 #####################
 if(UNIX)
-   set( viewer_platform_src "platform/Platform_Unix.cpp" )
+   set( viewer_platform_src "common/platform/Platform_Unix.cpp" )
 else()
-   set( viewer_platform_src "platform/Platform_Windows.cpp" )
+   set( viewer_platform_src "common/platform/Platform_Windows.cpp" )
 endif()
 file( GLOB common_src "common/*.cpp" "common/*.c" )
 set( common_src ${common_src} ${viewer_platform_src} )
@@ -56,24 +59,24 @@ set( common_src ${common_src} ${viewer_platform_src} )
 ### HOP GUI
 #####################
 file(GLOB hop_gui_src
-    "*.cpp"
+    "hop/*.cpp"
     "thirdparties/imgui/*.cpp"
 )
 
-add_executable(hop ${common_src} ${hop_gui_src} "front_ends/hopGUI.cpp" )
+add_executable(hop ${common_src} ${hop_gui_src} )
 target_compile_definitions( hop PUBLIC HOP_ENABLED )
 
 set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -DHOP_DEBUG -D${NOC_IMPL_DEFINE}")
 
-target_include_directories( hop SYSTEM PRIVATE ${ROOT_DIR} ${OPENGL_INCLUDE_DIRS} ${SDL2_INCLUDE_DIR} ${ROOT_DIR}/thirdparties )
-target_link_libraries( hop PUBLIC -lpthread ${HOP_LIBRT_FLAG} ${HOP_WINDOWS_LIB} ${OPENGL_LIBRARIES} ${SDL2_LIBRARY} -ldl )
+target_include_directories( hop SYSTEM PRIVATE ${ROOT_DIR} ${ROOT_DIR}/thirdparties ${OPENGL_INCLUDE_DIRS} ${SDL2_INCLUDE_DIR} ${PLATFORM_EXTRA_INCLUDES} )
+target_link_libraries( hop PUBLIC ${PLATFORM_LINK_FLAGS} ${HOP_WINDOWS_LIB} ${OPENGL_LIBRARIES} ${SDL2_LIBRARY} )
 
 #####################
 ### HOP CLI
 #####################
-add_executable( hopcli front_ends/hopCLI.cpp  ${common_src} )
+add_executable( hopcli hopcli/hopCLI.cpp  ${common_src} )
 target_include_directories( hopcli SYSTEM PRIVATE ${ROOT_DIR} )
-target_link_libraries( hopcli PUBLIC -lpthread ${HOP_LIBRT_FLAG} -ldl )
+target_link_libraries( hopcli PUBLIC ${PLATFORM_LINK_FLAGS} )
 target_compile_definitions( hopcli PUBLIC HOP_ENABLED )
 
 #####################
