+++ common/platform/Platform_Windows.cpp
