+++ hop/Cursor.cpp
@@ -1,4 +1,4 @@
-#include "Cursor.h"
+#include "hop/Cursor.h"
 
 #include <SDL.h>
