+++ hop/Lod.h
