+++ hop/Stats.h
