+++ hop/TraceStats.h
